Registry can store versions of questionnaires
I have changed the db model such that the most important information reg an survey is extracted and the rest is stored in a JSON blob. Additionally, the register now calculates a version number for every new addition to the db based on the number of previous instances of the same survey stored. Finally, it also stores the date that a survey was published, which is generated within the registry before the survey is saved.